SquashFS
Новый формат игровых архивов. Лучше сжатие, быстрее распаковка/упаковка, нет лимита по размеру архива. Архивы нового формата должны иметь расширение `.sq*`.
Для упаковки: https://github.com/pmq20/squashfuse/files/691217/sqfs43-win32.zip
```cmd
sqfs43-win32\mksquashfs.exe gamedata gamedata.sq_01 -comp lz4 -Xhc
pause
```
Для распаковки подойдёт https://github.com/mcmilk/7-Zip-zstd/releases/download/v24.09-v1.5.7-R1/7z24.09-zstd-x64.exe
-------------------------------------
Так же добавлена поддержка Anomaly-style архивов в подкаталогах. Дополнительные каталоги для поиска можно прописать в `fsgame.ltx`:
```ini
$app_arch_root$         = true|   false|  $fs_root$		  | game_archs\
```

+++ ogsr_engine/xrCore/file_stream_reader.cpp
@@ -1,12 +1,12 @@
 #include "stdafx.h"
 #include "file_stream_reader.h"
 
-void CFileStreamReader::construct(LPCSTR file_name, const u32& window_size)
+void CFileStreamReader::construct(LPCSTR file_name, const size_t& window_size)
 {
     m_file_handle = CreateFile(file_name, GENERIC_READ, FILE_SHARE_READ, 0, OPEN_EXISTING, 0, 0);
 
     VERIFY(m_file_handle != INVALID_HANDLE_VALUE);
-    u32 file_size = (u32)GetFileSize(m_file_handle, NULL);
+    size_t file_size = (size_t)GetFileSize(m_file_handle, NULL);
 
     HANDLE file_mapping_handle = CreateFileMapping(m_file_handle, 0, PAGE_READONLY, 0, 0, 0);
     VERIFY(file_mapping_handle != INVALID_HANDLE_VALUE);
